Update Arcgis and switch basemap implementation
The switch basemap of Arcgis only works in the last version of Arcgis so
implementing that functionality I had to update the arcgis plugin.
Previously we had a problem with the last version of Arcgis but it was
fixed by cleaning the 'hash' property  every time refreshed the map.

This resolve the issues #19 and #34